backend/requirements.txt
Django>=4.0
djangorestframework
djongo
python-dotenv
pytz
mongoengine
pymongo
bcrypt
djangorestframework-simplejwt
djangorestframework-jwt
django-cors-headers